Introduce typescript object model for workspace/page
This is an improvement because it encapsulates a lot of the bare
container/resource logic in an object model, but this still feels
awkward. There's also an outstanding bug where setting the name
doesn't update it in all places - we still need to use some sort of
react-aware update function to make sure updating the page name
changes it in all places.