Proof of concept of automerge working?
The basic approach here is to load a Slate document into an Automerge
doc and apply changes to the Automerge doc alongside the Slate
document.

The problem is that the Automerge doc gets very big very fast and is
not suitable for sending across the wire. It may be that we need to
build a custom server to handle collaborative editing and make the pod
data the "source of truth". For now I think it makes sense to just
stick with a basic patching workflow.