make concept ids base32-encoded encodeURIComponent'd strings
We want to use a really basic character set for ids, which show up in
URLs, thanks to inconsistent handling of various characters by various
systems that handle URIs generated by this system.

In particular, having a `%` seems to be problematic because some
systems double-decode them or don't decode them enough and `.` at the
end of a path name seems to cause some problems for the current Solid
implementation.